Bug fixes and consistency
