+++ src/Disqord.Core/Entities/Local/ApplicationCommands/Extensions/LocalSlashCommandOptionExtensions.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using Disqord.Models;
-using Disqord.Serialization.Json;
 
 namespace Disqord
 {
@@ -68,23 +65,5 @@ namespace Disqord
             option._channelTypes.AddRange(channelTypes);
             return option;
         }
-
-        public static ApplicationCommandOptionJsonModel ToModel(this LocalSlashCommandOption option, IJsonSerializer serializer)
-            => option == null ? null : new ApplicationCommandOptionJsonModel
-            {
-                Type = option.Type,
-                Name = option.Name,
-                Description = option.Description,
-                Required = option.IsRequired,
-                Choices = option.Choices.Select(x => x.ToModel(serializer)).ToArray(),
-                Options = option.Options.Select(x => x.ToModel(serializer)).ToArray()
-            };
-
-        public static ApplicationCommandOptionChoiceJsonModel ToModel(this LocalSlashCommandOptionChoice choice, IJsonSerializer serializer)
-            => choice == null ? null : new ApplicationCommandOptionChoiceJsonModel
-            {
-                Name = choice.Name,
-                Value = (IJsonValue)serializer.GetJsonNode(choice.Value)
-            };
     }
 }
